Машины и владельцы [#4744]

+++ db/scripts/008_ddl_create_table_car.sql
@@ -1,5 +1,5 @@
 CREATE TABLE car
 (
-    id        SERIAL PRIMARY KEY,
-    engine_id INT NOT NULL UNIQUE REFERENCES engine (id)
+    id        BIGSERIAL PRIMARY KEY,
+    engine_id BIGINT NOT NULL UNIQUE REFERENCES engine (id)
 );
